app/requirements.txt
requests
futures
google-cloud-pubsub
python-dotenv